+++ app/Services/Reply.php
@@ -692,8 +692,8 @@ class Reply extends ServiceBase
         $data['desc']           = shortname_to_unicode($reply->desc);
         $data['love_count']     = sCount::getLoveReplyNum($uid, $reply->id);
 
-        $data['is_follow']      = sFollow::checkRelationshipBetween($uid, $reply->uid);
-        $data['is_fan']         = sFollow::checkRelationshipBetween($reply->uid, $uid);
+        $data['is_follow']      = (int)sFollow::checkRelationshipBetween($uid, $reply->uid);
+        $data['is_fan']         = (int)sFollow::checkRelationshipBetween($reply->uid, $uid);
 
         $counts = cReplyCounts::get($reply->id);
         $data = array_merge($data, $counts);
